Fix buggy cast: Bug symptom:	hip_port_bindings_get() reports incorrect port binding information, leading to incorrect LSI handling. Bug cause:	When converting the port to look up from network to host byte order, it is stored and implicitely cast from in_port_t to uint8_t, losing the most significant byte of the port number. Bug fix:	Change the storage type of the port in host byte order from uint8_t to uint16_t to preserve its full numeric range.

+++ firewall/port_bindings.c
@@ -438,7 +438,7 @@ enum hip_port_binding hip_port_bindings_get(const uint8_t protocol,
     // check input parameters
     if (IPPROTO_TCP == protocol ||
         IPPROTO_UDP == protocol) {
-        const uint8_t port_hbo = ntohs(port);
+        const uint16_t port_hbo = ntohs(port);
 
         // Make sure we return (sort of) up-to-date information.
         // This is the one potentially slow operation here.
